MIC-701 CSIP overview

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppschallengesupportinterventionplanapi/model/CsipOverview.kt
@@ -2,12 +2,12 @@ package uk.gov.justice.digital.hmpps.hmppschallengesupportinterventionplanapi.mo
 
 data class CsipOverview(val counts: CsipCounts)
 data class CsipCounts(
-  val submittedReferrals: Int,
-  val pendingInvestigations: Int,
-  val awaitingDecisions: Int,
-  val pendingPlans: Int,
-  val open: Int,
-  val overdueReviews: Int,
+  val submittedReferrals: Long,
+  val pendingInvestigations: Long,
+  val awaitingDecisions: Long,
+  val pendingPlans: Long,
+  val open: Long,
+  val overdueReviews: Long,
 ) {
   companion object {
     val NONE = CsipCounts(0, 0, 0, 0, 0, 0)
